fix: restore Transformers 4 compatibility
